requirements.txt
certifi==2021.5.30
fastapi==0.66.0
pip==21.1.3
pydantic==1.8.2
setuptools==49.6.0
starlette==0.14.2
typing-extensions==3.10.0.0
wheel==0.36.2
requests~=2.26.0
python-dotenv~=0.18.0
numpy~=1.21.1
pandas~=1.3.0
uvicorn[standard]~=0.14.0
elasticsearch~=7.13.4
elasticsearch-dsl
PyYAML~=5.4.1